fix: split mappers
* Extract ChainedInput and ChainedMapper classes

* Fix input javajet

* Set data autopropagation to false in ComponentModel

* Fix processor javajet
